backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
google-generativeai
httpx
python-jose[cryptography]
bcrypt
prometheus-fastapi-instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp-proto-grpc